Fix ADC sampling issues in FPGA

+++ Software/ADCTest/SW4STM32/ADCTest/Application/User/FFTSampling.h
@@ -18,7 +18,7 @@ typedef enum {
 	FFT_WINDOW_MAX = 4,
 } fft_window_t;
 
-uint32_t FFT_take_sample(uint32_t points, fft_window_t window, uint32_t dummyFreq);
+float FFT_take_sample(uint32_t points, fft_window_t window, uint32_t dummyFreq);
 void fft_spi_test();
 
 #endif /* APPLICATION_USER_FFTSAMPLING_H_ */
